improve wikipedia pre-processing pipeline

+++ wikipedia_dumps_process/download_script.py
@@ -4,7 +4,9 @@ import os
 import urllib.request
 
 import requests
+from multiprocessing import Pool, cpu_count
 from tqdm import tqdm
+import json
 
 
 # code taken from https://stackoverflow.com/questions/15644964/python-progress-bar-and-downloads
@@ -16,6 +18,8 @@ class DownloadProgressBar(tqdm):
 
 
 def download_url(url, output_path):
+    print('URL:', url)
+    print('Output path:', output_path)
     with DownloadProgressBar(unit='B', unit_scale=True,
                              miniters=1, desc=url.split('/')[-1]) as t:
         urllib.request.urlretrieve(url, filename=output_path, reporthook=t.update_to)
@@ -33,8 +37,10 @@ if __name__ == '__main__':
                         required=True, help='output directory')
     parser.add_argument('--overwrite', action='store_true', default=False,
                         help='overwrite existing files')
+    parser.add_argument('--download_history', action='store_true', default=False,
+                        help='download revision history xml dump')
 
-    parser.set_defaults(overwrite=False)
+    parser.set_defaults(overwrite=False, download_history=False)
     args = parser.parse_args()
 
     # check if output directory exists
@@ -58,15 +64,6 @@ if __name__ == '__main__':
 
     url_sql_dump_3 = f"https://dumps.wikimedia.org/{args.lang}wiki/{args.date}/{args.lang}wiki-{args.date}-redirect.sql.gz"
     output_sql_path_3 = f"{args.output_dir}/{args.lang}wiki-{args.date}-redirect.sql.gz"
-
-    # check if files exist before downloading
-    if not os.path.exists(output_xml_path) or args.overwrite:
-        try:
-            download_url(url_xml_dump, output_xml_path)
-        except urllib.error.HTTPError:
-            print(f'Could not download {url_xml_dump}. This most likely means the dump is too large and the revision history is split into multiple files.')
-            print(f'Downloading json with revision history details. Individual revisions will be downloaded dynamically during processing.')
-            download_url(backup_json_dump, output_backup_path)
             
     if not os.path.exists(output_html_path) or args.overwrite:
         download_url(url_html_dump, output_html_path)
@@ -78,20 +75,45 @@ if __name__ == '__main__':
         download_url(url_sql_dump_3, output_sql_path_3)
 
     # extract the downloaded files (sql)
-    print(f'Extracting {output_sql_path_1}')
-    with gzip.open(output_sql_path_1, 'rb') as f:
-        file_content = f.read()
-        with open(output_sql_path_1[:-3], 'wb') as f_out:
-            f_out.write(file_content)
+    if not os.path.exists(output_sql_path_1[:-3]) or args.overwrite:
+        print(f'Extracting {output_sql_path_1}')
+        with gzip.open(output_sql_path_1, 'rb') as f:
+            file_content = f.read()
+            with open(output_sql_path_1[:-3], 'wb') as f_out:
+                f_out.write(file_content)
 
-    print(f'Extracting {output_sql_path_2}')
-    with gzip.open(output_sql_path_2, 'rb') as f:
-        file_content = f.read()
-        with open(output_sql_path_2[:-3], 'wb') as f_out:
-            f_out.write(file_content)
+    if not os.path.exists(output_sql_path_2[:-3]) or args.overwrite:
+        print(f'Extracting {output_sql_path_2}')
+        with gzip.open(output_sql_path_2, 'rb') as f:
+            file_content = f.read()
+            with open(output_sql_path_2[:-3], 'wb') as f_out:
+                f_out.write(file_content)
 
-    print(f'Extracting {output_sql_path_3}')
-    with gzip.open(output_sql_path_3, 'rb') as f:
-        file_content = f.read()
-        with open(output_sql_path_3[:-3], 'wb') as f_out:
-            f_out.write(file_content)
+    if not os.path.exists(output_sql_path_3[:-3]) or args.overwrite:
+        print(f'Extracting {output_sql_path_3}')
+        with gzip.open(output_sql_path_3, 'rb') as f:
+            file_content = f.read()
+            with open(output_sql_path_3[:-3], 'wb') as f_out:
+                f_out.write(file_content)
+            
+    if not args.download_history:
+        exit(0)
+    print('Downloading revision history xml dump')
+    # check if files exist before downloading
+    if not os.path.exists(output_xml_path) or args.overwrite:
+        try:
+            download_url(url_xml_dump, output_xml_path)
+        except urllib.error.HTTPError:
+            print(f'Could not download {url_xml_dump}. This most likely means the dump is too large and the revision history is split into multiple files.')
+            print(f'Downloading json with revision history details.')
+            download_url(backup_json_dump, output_backup_path)
+            print('Downloading all revision history files using multiprocessing.')
+            with open(output_backup_path, 'r') as f:
+                backup_json = json.load(f)
+            data = backup_json['jobs']['metahistorybz2dump']['files']
+            urls = [f"https://dumps.wikimedia.org{data[f]['url']}" for f in data]
+            paths = [f"{args.output_dir}/{f}" for f in data]
+            with Pool(3) as p:
+                p.starmap(download_url, zip(urls, paths))
+            
+
